Add MatrixHistogram to authentications tab
