+++ x-pack/plugins/security_solution/public/common/components/authentication/helpers.tsx
@@ -19,8 +19,15 @@ import { getRowItemDraggables } from '../tables/helpers';
 
 import * as i18n from './translations';
 import { HostDetailsLink, NetworkDetailsLink, UserDetailsLink } from '../links';
-import { AuthenticationsEdges } from '../../../../common/search_strategy';
+import { AuthenticationsEdges, MatrixHistogramType } from '../../../../common/search_strategy';
 import { AuthTableColumns } from './types';
+import {
+  MatrixHistogramConfigs,
+  MatrixHistogramMappingTypes,
+  MatrixHistogramOption,
+} from '../matrix_histogram/types';
+import { LensAttributes } from '../visualization_actions/types';
+import { authenticationLensAttributes } from '../visualization_actions/lens_attributes/common/authentication';
 
 export const getHostDetailsAuthenticationColumns = (usersEnabled: boolean): AuthTableColumns => [
   getUserColumn(usersEnabled),
@@ -240,4 +247,47 @@ const SUCCESS_COLUMN: Columns<AuthenticationsEdges, AuthenticationsEdges> = {
     );
   },
   width: '8%',
+};
+
+export const authenticationsStackByOptions: MatrixHistogramOption[] = [
+  {
+    text: 'event.outcome',
+    value: 'event.outcome',
+  },
+];
+const DEFAULT_STACK_BY = 'event.outcome';
+
+enum AuthenticationsMatrixDataGroup {
+  authenticationsSuccess = 'success',
+  authenticationsFailure = 'failure',
+}
+
+export enum ChartColors {
+  authenticationsSuccess = '#54B399',
+  authenticationsFailure = '#E7664C',
+}
+
+export const authenticationsMatrixDataMappingFields: MatrixHistogramMappingTypes = {
+  [AuthenticationsMatrixDataGroup.authenticationsSuccess]: {
+    key: AuthenticationsMatrixDataGroup.authenticationsSuccess,
+    value: null,
+    color: ChartColors.authenticationsSuccess,
+  },
+  [AuthenticationsMatrixDataGroup.authenticationsFailure]: {
+    key: AuthenticationsMatrixDataGroup.authenticationsFailure,
+    value: null,
+    color: ChartColors.authenticationsFailure,
+  },
+};
+
+export const histogramConfigs: MatrixHistogramConfigs = {
+  defaultStackByOption:
+    authenticationsStackByOptions.find((o) => o.text === DEFAULT_STACK_BY) ??
+    authenticationsStackByOptions[0],
+  errorMessage: i18n.ERROR_FETCHING_AUTHENTICATIONS_DATA,
+  histogramType: MatrixHistogramType.authentications,
+  mapping: authenticationsMatrixDataMappingFields,
+  stackByOptions: authenticationsStackByOptions,
+  title: i18n.NAVIGATION_AUTHENTICATIONS_TITLE,
+  lensAttributes: authenticationLensAttributes as LensAttributes,
 };
