+++ x-pack/plugins/security_solution/public/common/components/visualization_actions/lens_attributes/common/authentication.ts
