+++ x-pack/plugins/security_solution/public/hosts/pages/navigation/authentications_query_tab_body.tsx
@@ -7,58 +7,11 @@
 
 import React from 'react';
 import { HostsComponentsQueryProps } from './types';
-import {
-  MatrixHistogramOption,
-  MatrixHistogramMappingTypes,
-  MatrixHistogramConfigs,
-} from '../../../common/components/matrix_histogram/types';
 import { MatrixHistogram } from '../../../common/components/matrix_histogram';
-import { HostsKpiChartColors } from '../../components/kpi_hosts/types';
-import * as i18n from '../translations';
-import { MatrixHistogramType } from '../../../../common/search_strategy/security_solution';
-import { authenticationLensAttributes } from '../../../common/components/visualization_actions/lens_attributes/hosts/authentication';
-import { LensAttributes } from '../../../common/components/visualization_actions/types';
 import { AuthenticationsHostTable } from '../../../common/components/authentication/authentications_host_table';
+import { histogramConfigs } from '../../../common/components/authentication/helpers';
 
 const HISTOGRAM_QUERY_ID = 'authenticationsHistogramQuery';
-
-const authenticationsStackByOptions: MatrixHistogramOption[] = [
-  {
-    text: 'event.outcome',
-    value: 'event.outcome',
-  },
-];
-const DEFAULT_STACK_BY = 'event.outcome';
-
-enum AuthenticationsMatrixDataGroup {
-  authenticationsSuccess = 'success',
-  authenticationsFailure = 'failure',
-}
-
-export const authenticationsMatrixDataMappingFields: MatrixHistogramMappingTypes = {
-  [AuthenticationsMatrixDataGroup.authenticationsSuccess]: {
-    key: AuthenticationsMatrixDataGroup.authenticationsSuccess,
-    value: null,
-    color: HostsKpiChartColors.authenticationsSuccess,
-  },
-  [AuthenticationsMatrixDataGroup.authenticationsFailure]: {
-    key: AuthenticationsMatrixDataGroup.authenticationsFailure,
-    value: null,
-    color: HostsKpiChartColors.authenticationsFailure,
-  },
-};
-
-const histogramConfigs: MatrixHistogramConfigs = {
-  defaultStackByOption:
-    authenticationsStackByOptions.find((o) => o.text === DEFAULT_STACK_BY) ??
-    authenticationsStackByOptions[0],
-  errorMessage: i18n.ERROR_FETCHING_AUTHENTICATIONS_DATA,
-  histogramType: MatrixHistogramType.authentications,
-  mapping: authenticationsMatrixDataMappingFields,
-  stackByOptions: authenticationsStackByOptions,
-  title: i18n.NAVIGATION_AUTHENTICATIONS_TITLE,
-  lensAttributes: authenticationLensAttributes as LensAttributes,
-};
 
 const AuthenticationsQueryTabBodyComponent: React.FC<HostsComponentsQueryProps> = ({
   deleteQuery,
