+++ x-pack/plugins/security_solution/public/users/pages/navigation/authentications_query_tab_body.tsx
@@ -7,9 +7,12 @@
 
 import React from 'react';
 import { AuthenticationsUserTable } from '../../../common/components/authentication/authentications_user_table';
+import { histogramConfigs } from '../../../common/components/authentication/helpers';
 import { AuthenticationsUserTableProps } from '../../../common/components/authentication/types';
-
+import { MatrixHistogram } from '../../../common/components/matrix_histogram';
 export const ID = 'usersAuthenticationsQuery';
+
+const HISTOGRAM_QUERY_ID = 'usersAuthenticationsHistogramQuery';
 
 export const AuthenticationsQueryTabBody = ({
   endDate,
@@ -24,18 +27,30 @@ export const AuthenticationsQueryTabBody = ({
   userName,
 }: AuthenticationsUserTableProps) => {
   return (
-    <AuthenticationsUserTable
-      endDate={endDate}
-      filterQuery={filterQuery}
-      indexNames={indexNames}
-      setQuery={setQuery}
-      deleteQuery={deleteQuery}
-      startDate={startDate}
-      type={type}
-      skip={skip}
-      docValueFields={docValueFields}
-      userName={userName}
-    />
+    <>
+      <MatrixHistogram
+        endDate={endDate}
+        filterQuery={filterQuery}
+        id={HISTOGRAM_QUERY_ID}
+        indexNames={indexNames}
+        setQuery={setQuery}
+        startDate={startDate}
+        {...histogramConfigs}
+      />
+
+      <AuthenticationsUserTable
+        endDate={endDate}
+        filterQuery={filterQuery}
+        indexNames={indexNames}
+        setQuery={setQuery}
+        deleteQuery={deleteQuery}
+        startDate={startDate}
+        type={type}
+        skip={skip}
+        docValueFields={docValueFields}
+        userName={userName}
+      />
+    </>
   );
 };
